Projeto/requirements.txt
altgraph==0.17.3
asgiref==3.6.0
aspose-words==24.2.0
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
boto3==1.26.1
botocore==1.29.122
cairocffi==1.6.0
CairoSVG==2.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
comtypes==1.2.0
contourpy==1.1.0
cryptography==40.0.2
cssselect2==0.7.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.7
django-bootstrap4==23.1
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-extensions==3.1.5
django-filter==23.3
django-localflavor==3.1
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-restframework==0.0.1
django-session-timeout==0.1.0
django-stdimage==6.0.1
django-storages==1.13.1
django-svg-image-form-field==1.0.1
django-templated-mail==1.1.1
django-tenants==3.4.4
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djhtml==1.5.2
djoser==2.2.0
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.10.0
fonttools==4.42.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
isodate==0.6.1
isort==5.10.1
jedi==0.19.0
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keyboard==0.13.5
kiwisolver==1.4.5
lxml==4.9.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
MouseInfo==0.1.3
mss==9.0.1
mypy-extensions==1.0.0
nest-asyncio==1.5.7
numpy==1.25.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathspec==0.11.1
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.6
psycopg2-binary==2.9.6
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycep-correios==5.0.0
pycodestyle==2.10.0
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.16.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.6.0
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-stdnum==1.18
python3-openid==3.2.0
pytweening==1.0.7
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.1
pywinauto==0.6.8
pyzmq==25.1.1
regex==2023.3.23
requests==2.29.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
s3transfer==0.6.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
selenium==4.15.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
static3==0.7.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tinysvg==0.0.3
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
webdriver-manager==4.0.1
webencodings==0.5.1
whitenoise==6.2.0
wsproto==1.2.0
zeep==4.2.1
